requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.22.6
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
h11==0.14.0
idna==3.4
lazy-model==0.2.0
motor==3.3.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.8.0
uvicorn==0.23.2
